+++ pos_multi_session/static/src/js/pos_multi_session.js
@@ -933,7 +933,10 @@ odoo.define('pos_multi_session', function(require){
                     message: message,
                     dbname: session.db,
                     user_ID: self.pos.user.id
-                },{timeout:30000});
+                }, {
+                    shadow: true,
+                    timeout:30000
+                });
             };
             return send_it().fail(function (error, e) {
                 if (self.pos.debug){