+++ pos_multi_session/doc/changelog.rst
@@ -1,3 +1,15 @@
+`4.2.12`
+-------
+
+- **Fix:** Expected singleton error on adding POSes to a multisession
+
+`4.2.11`
+--------
+
+- **Fix:** Access error for poses created for different companies
+- **Fix:** Error during installation in multi-company mode
+- **Improvement:** Company rule  for multi-session model
+
 `4.2.10`
 --------